Move hex tests
